refactor : PointGroupFactory 클래스로 PointGroup 객체 생성 로직 이동
