+++ src/main/java/coordinate/domain/point/PointGroupFactory.java
@@ -1,7 +1,4 @@
-package coordinate.util;
-
-import coordinate.domain.point.Point;
-import coordinate.domain.point.PointGroup;
+package coordinate.domain.point;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -9,22 +6,15 @@ import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-public class CoordinateRepresentation {
+public class PointGroupFactory {
     private static final String ORDERED_PAIRS_SEPERATOR = "-";
     private static final Pattern orderedPairPattern = Pattern.compile("\\((.*),(.*)\\)");
 
-    public static PointGroup convertCoordinatePair(String orderedPairs) {
+    public static PointGroup create(String orderedPairs) {
         List<Point> points = new ArrayList<>();
         Arrays.asList(orderedPairs.split(ORDERED_PAIRS_SEPERATOR))
-                .forEach(x -> addNonOverlapPoint(points, convertCoordinate(x)));
+                .forEach(x -> points.add(convertCoordinate(x)));
         return new PointGroup(points);
-    }
-
-    private static void addNonOverlapPoint(List<Point> points, Point point) {
-        if (points.contains(point)) {
-            throw new IllegalArgumentException("중복된 점은 불가능합니다");
-        }
-        points.add(point);
     }
 
     private static Point convertCoordinate(String orderedPair) {
